Rename action_base -> action
