+++ bin/tools/templates/__action_template.h
@@ -2,7 +2,7 @@
 
 
 
-#include <fullscore/actions/action_base.h>
+#include <fullscore/actions/base.h>
