+++ include/fullscore/actions/delete_grid_column_action.h
@@ -2,7 +2,7 @@
 
 
 
-#include <fullscore/actions/action_base.h>
+#include <fullscore/actions/base.h>
