+++ include/fullscore/actions/move_cursor_up_action.h
@@ -2,7 +2,7 @@
 
 
 
-#include <fullscore/actions/action_base.h>
+#include <fullscore/actions/base.h>
