+++ include/fullscore/actions/set_command_mode_action.h
@@ -2,7 +2,7 @@
 
 
 
-#include <fullscore/actions/action_base.h>
+#include <fullscore/actions/base.h>
 
 #include <fullscore/widgets/grid_editor.h>
