+++ include/fullscore/actions/set_current_grid_editor_action.h
@@ -2,7 +2,7 @@
 
 
 
-#include <fullscore/actions/action_base.h>
+#include <fullscore/actions/base.h>
