+++ include/fullscore/actions/transforms/toggle_rest_action.h
@@ -3,7 +3,7 @@
 
 
 #include <vector>
-#include <fullscore/actions/action_base.h>
+#include <fullscore/actions/base.h>
